pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "gui-grounding"
version = "0.1.0"
description = "Cross-Website GUI Grounding with Verifiable Reward Optimization"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "Che Cheng"}]

dependencies = [
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "transformers>=4.40.0",
    "datasets>=2.18.0",
    "accelerate>=0.28.0",
    "peft>=0.10.0",
    "Pillow>=10.0.0",
    "pydantic>=2.5.0",
    "PyYAML>=6.0",
    "omegaconf>=2.3.0",
    "gradio>=4.20.0",
    "numpy>=1.24.0",
    "tqdm>=4.66.0",
    "wandb>=0.16.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.3.0",
    "black>=24.0.0",
    "isort>=5.13.0",
    "pre-commit>=3.6.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["gui_grounding"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["gui_grounding"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
